+++ project/test/test_event.gd
@@ -71,6 +71,16 @@ func test_event_environment() -> void:
 	assert_str(event.environment).is_equal("custom-environment")
 
 
+## SentryEvent.set_tag() should set tag to the specified value, remove_tag() should unset it.
+func test_event_tags() -> void:
+	var event := SentrySDK.create_event()
+	assert_str(event.get_tag("test_tag")).is_empty()
+	event.set_tag("test_tag", "test_value")
+	assert_str(event.get_tag("test_tag")).is_equal("test_value")
+	event.remove_tag("test_tag")
+	assert_str(event.get_tag("test_tag")).is_empty()
+
+
 ## SentrySDK.capture_event() should return a non-empty event ID, which must match the ID returned by the get_last_event_id() call.
 func test_capture_event() -> void:
 	var event := SentrySDK.create_event()